src/requirements.txt
playwright
playwright-stealth
gspread
google-auth
google-generativeai
python-dotenv
firebase-admin
